// admin-ui/src/app/component/shared/enum/base.enum.ts
export enum ServiceCategoryEnum {
  HostFlight = 1,
  HostExperience = 2,
  HostUniqueHome = 3,
  HostHotel = 4,
  HostTour = 5,
  HostBusiness = 6,
  HostTaxi = 7,
  HostUber = 8,
  CMS = 13,
  BookFlight = 21,
  BookExperience = 22,
  BookUniqueHome = 23,
  BookHotel = 24,
  BookTour = 25,
  BookBusiness = 26,
  BookTaxi = 27,
  BookUber = 28,
  Languages = 101,
  RoomType = 102,
  PropertyType = 103,
  TotalRooms = 104,
  TotalBeds = 105,
  TotalBathrooms = 106,
  TotalAdultGuest = 107,
  TotalChildGuest = 108,
  Amenities = 109,
  Highlights = 110,
  PlaceType = 111,
  BookingPreferences = 112,
  HouseRules = 113,
  CancellationPolicy = 114,
  ExperienceType = 115,
  BusinessKeywords = 116,
  VehicleType = 117,
  RoomOffers = 118,
  Neighbourhood = 119,
  PropertyFacilities = 120,
  FlightStoppage = 121,
  FlightFareType = 122,
  FlightAirlines = 123,
  FlightTicketClass = 124,
  TaxiType = 125,
  TaxiBookingType = 126,
  VendorType = 127,
  UserType = 128,
  UserStatus = 129,
  ServiceTypeStatus = 130,
  HostServiceStatus = 131,
  AttachmentStatus = 132,
  StayStatus = 133,
  TourItenaryStatus = 134,
  SightStatus = 135,
  FlightStatus = 136,
  TourDepartureStatus = 137,
  BookingStatus = 138,
  ReviewStatus = 139,
  AddStatus = 140,
  Fees = 141,
  Type = 142,
  Identity = 144,
  Sort = 153
}

export enum AttachmentType {
  Image = 158,
  Video = 159,
  Review = 160,
  Text = 198
}

export enum PaginationEnum {
  PageSize = 10
}

export enum DefaultRole {
  SuperUser = 1,
  LabOwner = 2,
  User = 3,
  Administrator = 4
}

export enum ControlType {
  TextBox = 1,
  CheckBox = 2,
  RadioButton = 3,
  Date = 4,
  Number = 5,
  DropDown = 6
}

export enum AssetTypeCategory {
  DocumentFolder = 1,
  DocumentStore = 2,
  LabPosition = 3,
  Employee = 4,
  Device = 5
}

export enum AssetType {
  Document = 1,
  Employee = 2,
  Device = 3
}

export enum ModuleCategory {
  DocumentFile = 3,
  DocumentFolder = 2,
  CheckList = 1,
  Employee = 4

}

export enum LabPosition {
  Director = 2,
  TechnicalConsultant = 3,
  ClinicalConsultant = 4,
  TestingPersonnel = 5,
  GeneralSupervisor = 6,
  TechnicalSupervisor = 7
}

export enum classification {
  Moderate = 1,
  HighComplexity = 2
}

export enum component {
  Dropdown,
  CheckboxWithAttachment,
  Attachment,
  RadioWithAttachment
}

export enum ServiceType {
  Experience = 2,
  UniqueHome = 3,
  Business = 6,
  Taxi = 7,
  Tour = 14
}


export enum LoginEnum {
  LIAMTRA = 'LIAMTRA',
  FacebookProvider = 'FACEBOOK',
  GoogleProvider = 'GOOGLE',
  AdminPanel = 'AdminPanel'
}